// src/app/modules/adzone-admin/models/admin-notification.model.ts
export interface NotificationMastermodel{
    actionId?: string;
    createdBy?: string;
    createdDate?: string;
    deliveredOn?: string;
    read?: boolean;
    message?: string;
    notificationId?: number;
    notificationModule?: string;
    orderId?: number;
    orderNumber?: number;
    orderOn?: string;
    orderStatus?: string;
    racfId?: string;
    redirectionId?: string;
    redirectionPage?: string;
    updatedBy?: string;
    updatedDate?: string;
    feedBack?: any;
    type?: any;
    documentDeletedBy?: any;
    documentId?:any;
    fileName?:any;
    documentFileName?:any;
    documentFileId?:any;
    documentExpiryDate?:any;
    approveDate?:any;

}